+++ database/column/float.go
@@ -39,14 +39,14 @@ func (f *Float) Row(stat *client.SessionDataSet, ptr bool) any {
 		if ptr {
 			return nil
 		}
-		return 0
+		return float64(0)
 	}
 	value, err := stat.GetFloat(f.name)
 	if err != nil {
 		if ptr {
 			return nil
 		}
-		return 0
+		return float64(0)
 	}
 	if ptr {
 		return &value
